Swagger added. refactor validations and entities. add @RequiredArgsConstructor annotation. change id -> Long.

+++ src/main/java/com/example/producttestapi/repository/CartItemRepo.java
@@ -5,7 +5,7 @@ import com.example.producttestapi.entity.Product;
 import com.example.producttestapi.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartItemRepo extends JpaRepository<CartItem , Integer> {
+public interface CartItemRepo extends JpaRepository<CartItem , Long> {
 
     CartItem findByProductAndCart_User(Product product, User user);
 }
