+++ src/main/java/com/example/producttestapi/repository/CartRepo.java
@@ -4,6 +4,6 @@ import com.example.producttestapi.entity.Cart;
 import com.example.producttestapi.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartRepo extends JpaRepository<Cart, Integer> {
+public interface CartRepo extends JpaRepository<Cart, Long> {
     Cart findByUser(User user);
 }
