+++ src/main/java/com/example/producttestapi/repository/CategoryRepo.java
@@ -6,10 +6,10 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import java.util.List;
-public interface CategoryRepo extends JpaRepository<Category, Integer> {
+public interface CategoryRepo extends JpaRepository<Category, Long> {
     @Query("SELECT c FROM Category c WHERE c.parentCategory IS null")
     List<Category> getAllMainCategories();
 
     @Query("SELECT c FROM Category c WHERE c.parentCategory.id = :categoryId")
-    List<Category> getCategoryChildren(@Param("categoryId") Integer categoryId);
+    List<Category> getCategoryChildren(@Param("categoryId") Long categoryId);
 }
