+++ src/main/java/com/example/producttestapi/repository/UserRepo.java
@@ -3,9 +3,9 @@ package com.example.producttestapi.repository;
 import com.example.producttestapi.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-    public interface UserRepo extends JpaRepository<User, Integer> {
+    public interface UserRepo extends JpaRepository<User, Long> {
     User findByEmail(String  email);
-    User getUserById(int id);
+    User getUserById(Long id);
     User findAllByEmail(String email);
-    void deleteById(int id);
+    void deleteById(Long id);
 }
