+++ src/main/java/com/example/producttestapi/repository/VoucherRepo.java
@@ -3,6 +3,6 @@ package com.example.producttestapi.repository;
 import com.example.producttestapi.entity.Voucher;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VoucherRepo extends JpaRepository<Voucher, Integer> {
+public interface VoucherRepo extends JpaRepository<Voucher, Long> {
     Voucher findByCode(String code);
 }
