+++ src/main/java/com/example/producttestapi/service/CartService.java
@@ -3,7 +3,7 @@ package com.example.producttestapi.service;
 import com.example.producttestapi.dto.CartDto;
 
 public interface CartService {
-     void deleteCart(int id) ;
+     void deleteCart(Long id) ;
 
     CartDto getCart();
 }
